requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
coverage==7.8.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
factory_boy==3.3.3
Faker==37.1.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytest-factoryboy==2.7.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
ruff==0.11.6
setuptools==78.0.2
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
wheel==0.45.1
